feat: Detoggle sprint if at zero stam for elapsed time
This can come in handy in situations involving low stamina and/or
debuffs like rain and cold are active, or no rested buff. Stamina
disappears quickly and is slow to regenerate. The first impulse is to
override forward to escape danger or attack, and doing so while sprint
is toggled will keep stamina at zero for as long as the Forward
key/button is held. This isn't a great experience, so this feature can
help